Refactor type definitions to avoid runtime imports
