+++ src/app/api/generate-character/route.ts
@@ -2,7 +2,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import { createAPIErrorResponse } from '@/lib/utils/errorUtils';
 import { generateCharacter } from '@/lib/ai/characterGenerator';
 import { World } from '@/types/world.types';
-import { validateWorld } from '@/types/type-guards';
+import { validateWorld } from '@/lib/utils/typeGuards';
 
 export async function POST(request: NextRequest) {
   try {
