+++ src/types/genre.types.ts
@@ -1,0 +1,15 @@
+// src/types/genre.types.ts
+
+/**
+ * Supported genre values.
+ */
+export type GenreValue =
+  | 'fantasy'
+  | 'sci-fi'
+  | 'modern'
+  | 'historical'
+  | 'horror'
+  | 'mystery'
+  | 'western'
+  | 'cyberpunk'
+  | 'other';
