+++ src/types/skill-difficulty.types.ts
@@ -1,0 +1,6 @@
+// src/types/skill-difficulty.types.ts
+
+/**
+ * Difficulty levels for world/skill definitions.
+ */
+export type SkillDifficulty = 'easy' | 'medium' | 'hard';
